This is the RESTAURANT-RECIPE project first commit
First of all we have the main file and the main home page, the index.html file
The second is the css file which applies styling to the main page.
Finally we have the three html files which are pages or links to the recipes
from the first page (index.html).